Add alternative errnoexception with user specified errno Minor cleanups

+++ tests/TestProcess.cpp
@@ -19,8 +19,7 @@ void TestProcess::tearDown()
 
 void TestProcess::Test()
 {
-
-	bool status;
+	bool status = false;
 	string reply;
 
 	logg << Logger::Debug << "Testprocess start" <<lend;
@@ -35,10 +34,15 @@ void TestProcess::Test()
 
 	CPPUNIT_ASSERT_NO_THROW(Utils::Process::Spawn( "/bin/ls" ) );
 
-	File::Delete("/tmp/echotest");
+	if( File::FileExists("/tmp/echotest"))
+	{
+		File::Delete("/tmp/echotest");
+	}
+
 	CPPUNIT_ASSERT_NO_THROW(Utils::Process::Spawn( "/usr/bin/touch",{"/tmp/echotest"} ) );
 	sleep(1);
 	CPPUNIT_ASSERT(File::FileExists("/tmp/echotest"));
 
 	logg << Logger::Debug << "TestProcess complete" <<lend;
+
 }
